Boostrap implementation (#12)
* Implement boostrap static data

* Expose GetById , GetByGW and GetAll for FPLFixture

* Complete boostrap model

* Enhanced models

* Add more Models and properties to the model folder

* Resolve cyclic reference in player position in Boostrap call of FplElement

* Add load profile on login to injection option
Introduce ILoadProfile to load profile and event

* Add show log to see if there is error or which part is delaying

Guide against infinite wait , and attempt retry instead

* Method to remove cookie even after initial page Login cookie

* refresh

* upadte the sample to include detail and force refresh

* Implement My-team, Classic League & H2H

* Add selling price and purchase price

* Add Selling and Cost price to FplPick

* Add Method to get live player data to PlayerClient

+++ src/FantasyPremierLeague/Models/Teams/FplEntryHistory.cs
@@ -41,13 +41,13 @@ public sealed class FplEntryHistory
     /// Gets or sets the bank.
     /// </summary>
     [JsonPropertyName("bank")]
-    public int Bank { get; set; }
+    public float Bank { get; set; }
 
     /// <summary>
-    /// Gets or sets the value.
+    /// Gets or sets the squad value.
     /// </summary>
     [JsonPropertyName("value")]
-    public int Value { get; set; }
+    public float Value { get; set; }
 
     /// <summary>
     /// Gets or sets the event transfers.
